rename components folder
