+++ ozaria/site/components/teacher-dashboard/modals/ModalEditClassV2.vue
@@ -86,8 +86,8 @@ import GoogleClassroomHandler from 'core/social-handlers/GoogleClassroomHandler'
 import LmsRosterImportHandler from 'core/social-handlers/LmsRosterImportHandler'
 
 import Modal from '../../common/Modal'
-import PageFirst from './ModalEditClass/PageFirst'
-import PageSecond from './ModalEditClass/PageSecond'
+import PageFirst from './edit-class-components-v2/PageFirst'
+import PageSecond from './edit-class-components-v2/PageSecond'
 import PurpleButton from '../common/buttons/PurpleButton'
 import PurpleSecondaryButton from '../common/buttons/PurpleSecondaryButton'
 import { COMPONENT_NAMES } from 'ozaria/site/components/teacher-dashboard/common/constants.js'
