+++ ozaria/site/components/teacher-dashboard/modals/edit-class-components-v2/CourseSelect.vue
